// packages/codex-orchestration-cli/hooks/generated/session-start-input.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type NullableString = string | null;

export interface SessionStartCommandInput {
  cwd: string;
  hook_event_name: "SessionStart";
  model: string;
  permission_mode: "default" | "acceptEdits" | "plan" | "dontAsk" | "bypassPermissions";
  session_id: string;
  source: "startup" | "resume" | "clear" | "compact";
  transcript_path: NullableString;
}