// packages/codex-orchestration-cli/hooks/generated/user-prompt-submit-output.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export type BlockDecisionWire = "block";

export interface UserPromptSubmitCommandOutput {
  continue?: boolean;
  decision?: BlockDecisionWire;
  hookSpecificOutput?: UserPromptSubmitHookSpecificOutputWire;
  reason?: string;
  stopReason?: string;
  suppressOutput?: boolean;
  systemMessage?: string;
}
export interface UserPromptSubmitHookSpecificOutputWire {
  additionalContext?: string;
  hookEventName: "UserPromptSubmit";
}